sync_positions/requirements.txt

requests
redis
pytz
mysql-connector-python